move metric specs to resource_slo
